Pull up method to abstract base class
