+++ sso/provider/pom.xml
@@ -34,5 +34,10 @@
 			<groupId>org.eclipse.kapua</groupId>
 			<artifactId>kapua-commons</artifactId>
 		</dependency>
+
+		<dependency>
+			<groupId>org.apache.httpcomponents</groupId>
+			<artifactId>httpclient</artifactId>
+		</dependency>
 	</dependencies>
 </project>
